+++ xarray/datatree_/datatree/__init__.py
@@ -6,7 +6,7 @@ from pkg_resources import DistributionNotFound, get_distribution
 # import public API
 from .datatree import DataTree
 from .io import open_datatree
-from .mapping import map_over_subtree
+from .mapping import TreeIsomorphismError, map_over_subtree
 
 try:
     __version__ = get_distribution(__name__).version